Fix euw login

+++ src/main/java/com/gvaneyck/rtmp/LoLRTMPSClient.java
@@ -471,7 +471,7 @@ public class LoLRTMPSClient extends RTMPSClient
 
         // Handle login queue
         if (!result.containsKey("token")) {
-            long node = result.getLong("node"); // Our login queue ID
+            int node = result.getInt("node"); // Our login queue ID
             String nodeStr = "" + node;
             String champ = result.getString("champ"); // The name of our login
                                                       // queue
